refactor: move db-specific functions out of utils
It was causing circular dependency; utils package should have have
generic reusable functions, not package-specific reusable functions.
